+++ ts/Series/Venn/VennPointOptions.d.ts
@@ -1,0 +1,38 @@
+/* *
+ *
+ *  Experimental Highcharts module which enables visualization of a Venn
+ *  diagram.
+ *
+ *  (c) 2016-2020 Highsoft AS
+ *  Authors: Jon Arild Nygard
+ *
+ *  Layout algorithm by Ben Frederickson:
+ *  https://www.benfrederickson.com/better-venn-diagrams/
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type ScatterPointOptions from '../Scatter/ScatterPointOptions';
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+export interface VennPointOptions extends ScatterPointOptions {
+    name?: string;
+    sets?: Array<string>;
+    value?: number;
+}
+
+export default VennPointOptions;
